+++ Basic/DistributedAuthoritySocialHub/Assets/Scripts/Gameplay/ISpawnable.cs
@@ -1,0 +1,15 @@
+using Unity.Netcode;
+using UnityEngine;
+
+namespace Unity.Multiplayer.Samples.SocialHub.Gameplay
+{
+    interface ISpawnable
+    {
+        NetworkObject NetworkObject
+        {
+            get;
+        }
+
+        void Init(SessionOwnerNetworkObjectSpawner networkObjectSpawner);
+    }
+}