+++ Experimental/DistributedAuthoritySample/Assets/Scripts/Physics/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-using System.Runtime.CompilerServices;
-[assembly: InternalsVisibleTo("Unity.Multiplayer.Samples.SocialHub.Player")]
-[assembly: InternalsVisibleTo("Unity.Multiplayer.Samples.SocialHub.Gameplay")]
-[assembly: InternalsVisibleTo("Unity.Multiplayer.Samples.SocialHub.Editor")]